Low: crmd: Moved to common function.
