+++ crmd/notify.c
@@ -85,218 +85,6 @@ static const char *notify_keys[][3] =
 };
 
 /*
- * higher accuracy time stuff to be generalized and moved to
- * e.g. lib/common/utils.c|iso8601.c
- */
-
-#include <time.h>
-#include <sys/time.h>
-#include <ctype.h>
-#include <crm/common/iso8601.h>
-
-typedef struct crm_time_us crm_time_hr_t;
-crm_time_hr_t *crm_time_hr_convert(crm_time_hr_t *target, crm_time_t *dt);
-void crm_time_set_hr_dt(crm_time_t *target, crm_time_hr_t *hr_dt);
-crm_time_hr_t *crm_time_timeval_hr_convert(crm_time_hr_t *target,
-                                           struct timeval *tv);
-crm_time_hr_t *crm_time_hr_new(const char *date_time);
-void crm_time_hr_free(crm_time_hr_t * hr_dt);
-char *crm_time_format_hr(const char *format, crm_time_hr_t * hr_dt);
-
-crm_time_t *parse_date(const char *date_str); /* in iso8601.c global but
-                                                 not in header */
-
-struct crm_time_us {
-    int years;
-    int months;                 /* Only for durations */
-    int days;
-    int seconds;
-    int offset;                 /* Seconds */
-    bool duration;
-    int useconds;
-};
-
-struct crm_time_s {
-    int years;
-    int months;                 /* Only for durations */
-    int days;
-    int seconds;
-    int offset;                 /* Seconds */
-    bool duration;
-};
-
-
-
-static void
-ha_get_tm_time( struct tm *target, crm_time_t *source)
-{
-    *target = (struct tm) {
-        .tm_year = source->years - 1900,
-        .tm_yday = source->days - 1,
-        .tm_sec = source->seconds % 60,
-        .tm_min = ( source->seconds / 60 ) % 60,
-        .tm_hour = source->seconds / 60 / 60,
-
-#if defined(HAVE_STRUCT_TM_TM_GMTOFF)
-        .tm_gmtoff = source->offset
-#endif
-    };
-}
-
-crm_time_hr_t *
-crm_time_hr_convert(crm_time_hr_t *target, crm_time_t *dt)
-{
-    crm_time_hr_t *hr_dt = NULL;
-
-    if (dt) {
-        hr_dt = target?target:calloc(1, sizeof(crm_time_hr_t));
-        if (hr_dt) {
-            *hr_dt = (crm_time_hr_t) {
-                .years = dt->years,
-                .months = dt->months,
-                .days = dt->days,
-                .seconds = dt->seconds,
-                .offset = dt->offset,
-                .duration = dt->duration
-            };
-        }
-    }
-
-    return hr_dt;
-}
-
-void
-crm_time_set_hr_dt(crm_time_t *target, crm_time_hr_t *hr_dt)
-{
-    CRM_ASSERT((hr_dt) && (target));
-    *target = (crm_time_t) {
-        .years = hr_dt->years,
-        .months = hr_dt->months,
-        .days = hr_dt->days,
-        .seconds = hr_dt->seconds,
-        .offset = hr_dt->offset,
-        .duration = hr_dt->duration
-    };
-}
-
-crm_time_hr_t *
-crm_time_timeval_hr_convert(crm_time_hr_t *target, struct timeval *tv)
-{
-    crm_time_t dt;
-    crm_time_hr_t *ret;
-
-    crm_time_set_timet(&dt, &tv->tv_sec);
-    ret = crm_time_hr_convert(target, &dt);
-    if (ret) {
-        ret->useconds = tv->tv_usec;
-    }
-    return ret;
-}
-
-crm_time_hr_t *
-crm_time_hr_new(const char *date_time)
-{
-    crm_time_hr_t *hr_dt = NULL;
-    struct timeval tv_now;
-
-    if (!date_time) {
-        if (gettimeofday(&tv_now, NULL) == 0) {
-            hr_dt = crm_time_timeval_hr_convert(NULL, &tv_now);
-        }
-    } else {
-        crm_time_t *dt;
-
-        dt = parse_date(date_time);
-        hr_dt = crm_time_hr_convert(NULL, dt);
-        crm_time_free(dt);
-    }
-    return hr_dt;
-}
-
-void
-crm_time_hr_free(crm_time_hr_t * hr_dt)
-{
-    free(hr_dt);
-}
-
-char *
-crm_time_format_hr(const char *format, crm_time_hr_t * hr_dt)
-{
-    const char *mark_s;
-    int max = 128, scanned_pos = 0, printed_pos = 0, fmt_pos = 0,
-        date_len = 0, nano_digits = 0, fmt_len;
-    char nano_s[10], date_s[max+1], nanofmt_s[5] = "%", *tmp_fmt_s;
-    struct tm tm;
-    crm_time_t dt;
-
-    if (!format) {
-        return NULL;
-    }
-    crm_time_set_hr_dt(&dt, hr_dt);
-    ha_get_tm_time(&tm, &dt);
-    sprintf(nano_s, "%06d000", hr_dt->useconds);
-
-    while ((format[scanned_pos]) != '\0') {
-        fmt_len = 0;
-        mark_s = strchr(&format[scanned_pos], '%');
-        if (mark_s) {
-            fmt_pos = mark_s - format;
-            fmt_len = 1;
-            while ((format[fmt_pos+fmt_len] != '\0') &&
-                (format[fmt_pos+fmt_len] >= '0') &&
-                (format[fmt_pos+fmt_len] <= '9')) {
-                fmt_len++;
-            }
-            scanned_pos = fmt_pos + fmt_len + 1;
-            if (format[fmt_pos+fmt_len] == 'N') {
-                nano_digits = atoi(&format[fmt_pos+1]);
-                nano_digits = (nano_digits > 6)?6:nano_digits;
-                nano_digits = (nano_digits < 0)?0:nano_digits;
-                sprintf(&nanofmt_s[1], ".%ds", nano_digits);
-            } else {
-                if (format[scanned_pos] != '\0') {
-                    continue;
-                }
-                fmt_pos = scanned_pos; /* print till end */
-            }
-        } else {
-            scanned_pos = strlen(format);
-            fmt_pos = scanned_pos; /* print till end */
-        }
-        tmp_fmt_s = strndup(&format[printed_pos], fmt_pos - printed_pos);
-#ifdef GCC_FORMAT_NONLITERAL_CHECKING_ENABLED
-#pragma GCC diagnostic push
-#pragma GCC diagnostic ignored "-Wformat-nonliteral"
-#endif
-        date_len += strftime(&date_s[date_len], max-date_len, tmp_fmt_s, &tm);
-#ifdef GCC_FORMAT_NONLITERAL_CHECKING_ENABLED
-#pragma GCC diagnostic pop
-#endif
-        printed_pos = scanned_pos;
-        free(tmp_fmt_s);
-        if (nano_digits) {
-#ifdef GCC_FORMAT_NONLITERAL_CHECKING_ENABLED
-#pragma GCC diagnostic push
-#pragma GCC diagnostic ignored "-Wformat-nonliteral"
-#endif
-            date_len += snprintf(&date_s[date_len], max-date_len,
-                                 nanofmt_s, nano_s);
-#ifdef GCC_FORMAT_NONLITERAL_CHECKING_ENABLED
-#pragma GCC diagnostic pop
-#endif
-            nano_digits = 0;
-        }
-    }
-
-    return (date_len == 0)?NULL:strdup(date_s);
-}
-
-/*
- * end of possibly generic time-handling stuff
- */
-
-
-/*
  * syncronize local data with cib
  */
